# backend/src/main/resources/openapi/anzol.yaml
openapi: 3.1.0
info:
  title: Anzol API
  version: 0.3.1
  description: >-
    Webhook inspector. A token is a capture URL (`/{tokenId}`) that records every HTTP request it receives and
    answers with its configured response or with the first matching response rule. The `/token/...` routes manage
    a URL and read what it captured.


    Request bodies are JSON objects; unknown fields are ignored. Routes that keep the legacy input model
    (`POST`/`PUT /token`, `unlock`, `keys`, `share`, `e2ee-lab`) also accept a form body, and query parameters
    are merged with the body. Errors are JSON for API clients (`Accept: application/json`, a JSON `Content-Type`
    or `X-Requested-With: XMLHttpRequest`) and an HTML page otherwise. Management routes check `Host` and, on
    state-changing methods, `Origin` against the allowed hosts (403 `{"error": "..."}`).
servers:
  - url: /
...
paths:
...
  /token/{tokenId}/requests:
    parameters:
      - $ref: '#/components/parameters/TokenId'
    get:
      tags: [Requests]
      operationId: listRequests
      summary: List the messages of a URL
      description: >-
        Paginated. With `after`, the messages with a greater `seq`, oldest first, up to `per_page` (`page` and
        `sorting` are ignored). Validation runs before the URL lookup.
      security:
        - {}
        - readSecret: []
        - accessCookie: []
      parameters:
        - name: page
          in: query
          schema: {type: integer, default: 1}
        - name: per_page
          in: query
          schema: {type: integer, default: 50}
        - name: sorting
          in: query
          description: Anything other than `newest` is `oldest`.
          schema: {type: string, enum: [oldest, newest], default: oldest}
        - name: after
          in: query
          description: A `seq`; only messages after it.
          schema: {type: integer, format: int64, minimum: 0}
      responses:
        '200':
          description: A page of messages.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestPage'
        '302':
          $ref: '#/components/responses/ValidationRedirect'
        '401':
          $ref: '#/components/responses/Protected'
        '410':
          $ref: '#/components/responses/Gone'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManySecretAttempts'
...
  /token/{tokenId}/requests/search:
    parameters:
      - $ref: '#/components/parameters/TokenId'
    post:
      tags: [Search]
      operationId: searchRequests
      summary: Search the messages
      description: >-
        By text and by a rule `match`, paginated like the listing (`total` is how many match). An empty body is
        `{}`.
      security:
        - {}
        - readSecret: []
        - accessCookie: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: A page of matching messages.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestPage'
        '401':
          $ref: '#/components/responses/Protected'
        '410':
          $ref: '#/components/responses/Gone'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManySecretAttempts'
...
components:
  securitySchemes:
    readSecret:
      type: apiKey
      in: header
      name: X-Anzol-Secret
      description: >-
        The read secret of the URL. Only required when the URL is protected (`protected: true`); other URLs
        ignore it.
    accessCookie:
      type: apiKey
      in: cookie
      name: anzol_access
      description: >-
        The unlock cookie set by `POST /token/{tokenId}/unlock`. Only required when the URL is protected.
  parameters:
    TokenId:
      name: tokenId
      in: path
      required: true
      description: The URL uuid (lowercase).
      schema:
        type: string
        format: uuid
...
  headers:
    RetryAfter:
      description: Seconds to wait before trying again.
      schema:
        type: integer
...
  responses:
...
    ValidationError:
      description: Validation failed. Keys are field paths; values are the messages.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrors'
          example:
            e2ee.path: [The e2ee.path field is required.]
    ValidationRedirect:
      description: >-
        Validation failed on a request that neither sends JSON nor accepts it (a form): redirect to the
        `Referer`, or to `/`.
      headers:
        Location:
          schema:
            type: string
...
    Protected:
      description: The URL is protected and the request has neither the read secret nor a valid unlock cookie.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProtectedError'
          example: {error: This URL is protected, protected: true}
    Gone:
      description: The URL does not exist (never did, expired or was deleted).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            error: {message: Token not found, id: null}
        text/html:
          schema:
            type: string
...
    TooManySecretAttempts:
      description: Too many wrong read secrets for this URL in the last minute.
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SimpleError'
          example: {error: Too many wrong secrets for this URL; try again later}
...
  schemas:
    Timestamp:
      type: string
      description: UTC, `yyyy-MM-dd HH:mm:ss`.
      pattern: '^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$'
      examples: ['2026-09-26 14:02:07']
...
    ValidationErrors:
      type: object
      additionalProperties:
        type: array
        items:
          type: string
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          const: false
        error:
          type: object
          required: [message, id]
          properties:
            message:
              type: string
            id:
              type: 'null'
    SimpleError:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ProtectedError:
      type: object
      required: [error, protected]
      properties:
        error:
          const: This URL is protected
        protected:
          const: true
...
    MessageFields:
      type: object
      description: The fields every message read has.
      required:
        - uuid
        - ip
        - hostname
        - method
        - user_agent
        - content
        - query
        - headers
        - url
        - created_at
        - updated_at
        - rule
        - near_miss
        - signature
        - schema
        - response
        - decryption
        - seq
      properties:
        uuid:
          type: string
          format: uuid
        ip:
          type: [string, 'null']
        hostname:
          type: string
        method:
          type: string
        user_agent:
          type: [string, 'null']
        content:
          type: string
          description: The body as captured.
        query:
          description: The query parameters, a PHP array (object or list).
          type: [object, array, 'null']
        headers:
          type: object
          description: Lowercase header name to its values.
          additionalProperties:
            type: array
            items:
              type: string
        url:
          type: string
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
        request:
          description: >-
            The parsed form fields; absent on JSON requests, `null` without fields.
          type: [object, array, 'null']
        rule:
          description: The rule that answered.
          oneOf:
            - $ref: '#/components/schemas/RuleRef'
            - type: 'null'
        near_miss:
          description: With no rule matching, the enabled rule closest to matching.
          oneOf:
            - $ref: '#/components/schemas/NearMiss'
            - type: 'null'
        signature:
          oneOf:
            - $ref: '#/components/schemas/SignatureResult'
            - type: 'null'
        schema:
          oneOf:
            - $ref: '#/components/schemas/SchemaResult'
            - type: 'null'
        response:
          oneOf:
            - $ref: '#/components/schemas/RecordedResponse'
            - type: 'null'
        decryption:
          oneOf:
            - $ref: '#/components/schemas/Decryption'
            - type: 'null'
        seq:
          type: integer
          format: int64
          description: Strictly increasing per URL in storage order; never reused.
    Message:
      description: A captured request.
      allOf:
        - $ref: '#/components/schemas/MessageFields'
        - type: object
          required: [token_id]
          properties:
            token_id:
              type: string
              format: uuid
            decrypted:
              description: >-
                The opened attribute (the JWS `data` claim); only when `decryption.state` is `valid`. Stored in
                plaintext in Redis with the message: the read secret guards reading it through the API, not the
                storage.
...
    RuleRef:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    NearMiss:
      type: object
      required: [id, name, failed, conditions]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        failed:
          type: array
          description: One phrase per failed condition.
          items:
            type: string
        conditions:
          description: The key of each failed condition, in the order of `failed`; `null` on old messages.
          type: [array, 'null']
          items:
            type: string
    SignatureResult:
      type: object
      required: [provider, valid, reason]
      properties:
        provider:
          type: string
        valid:
          type: boolean
        reason:
          type: [string, 'null']
          examples: [signature mismatch]
    SchemaResult:
      type: object
      required: [valid, errors]
      properties:
        valid:
          type: boolean
        errors:
          type: array
          maxItems: 20
          items:
            $ref: '#/components/schemas/SchemaError'
    SchemaError:
      type: object
      required: [path, message]
      properties:
        path:
          type: string
          description: JSON Pointer in the instance; `""` is the root.
        message:
          type: string
    RecordedResponse:
      type: object
      description: What the capture answered, the status or the rule fault that dropped the connection.
      properties:
        status:
          type: integer
        fault:
          $ref: '#/components/schemas/Fault'
    Decryption:
      type: object
      required: [state, kid, signature_kid, reason, jti, duplicate_of, kid_deleted_at, aud]
      properties:
        state:
          $ref: '#/components/schemas/DecryptionState'
        kid:
          type: [string, 'null']
          description: The encryption key id, when read.
        signature_kid:
          type: [string, 'null']
          description: The trusted signer that signed, when read.
        reason:
          type: [string, 'null']
          description: Why it failed, e.g. `hmac_failed`, `downgrade`, `decrypt_failed`, `aud_mismatch`.
        jti:
          type: [string, 'null']
        duplicate_of:
          description: On a redelivery of a decrypted `jti`, the first message.
          type: [string, 'null']
          format: uuid
        kid_deleted_at:
          description: >-
            With `unknown_kid` or `decrypt_failed`, when this URL deleted an encryption key with this `kid` (UTC,
            `yyyy-MM-dd HH:mm:ss`): the sender encrypted to a deleted key, or to one deleted and recreated. `null` when
            the URL has no record of it (it keeps its last 20 deleted keys) and in messages recorded before the field.
          type: [string, 'null']
        aud:
          description: >-
            The `aud` claim of the JWS once its signature verified, as a list (a single string becomes a list of one;
            non-string values are left out). Text written by the sender, shown as data: at most 5 values of at most
            256 characters, with control and format characters replaced by U+FFFD. `null` when decryption stopped
            before the JWS signature was verified, and in messages recorded before the field.
          type: [array, 'null']
          items:
            type: string
    DecryptionState:
      enum: [valid, invalid, unknown_kid, absent]
    RequestPage:
      type: object
      required: [data, total, per_page, current_page, is_last_page, from, to]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        total:
          type: integer
          format: int64
        per_page:
          type: integer
        current_page:
          type: integer
        is_last_page:
          type: boolean
        from:
          type: integer
        to:
          type: integer
...
    RuleMatch:
      type: object
      description: Conditions, all required (AND). No condition matches everything.
      properties:
        method:
          type: array
          description: Any of these methods (case-insensitive); empty matches any.
          items:
            type: string
        path:
          description: On the path after the uuid.
          oneOf:
            - $ref: '#/components/schemas/PathMatcher'
            - type: 'null'
        query:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/FieldMatcher'
        headers:
          type: object
          description: Header names are case-insensitive.
          additionalProperties:
            $ref: '#/components/schemas/FieldMatcher'
        body:
          type: array
          items:
            $ref: '#/components/schemas/BodyMatcher'
        signature:
          enum: [valid, invalid, absent]
        schema:
          enum: [valid, invalid]
        decryption:
          $ref: '#/components/schemas/DecryptionState'
    PathMatcher:
      description: Exactly one operator.
      oneOf:
        - type: object
          required: [equals]
          properties:
            equals: {type: string}
        - type: object
          required: [prefix]
          properties:
            prefix: {type: string}
        - type: object
          required: [regex]
          properties:
            regex:
              type: string
              description: Java regex matching the whole value.
    FieldMatcher:
      description: Exactly one operator.
      oneOf:
        - type: object
          required: [equals]
          properties:
            equals: {type: string}
        - type: object
          required: [contains]
          properties:
            contains: {type: string}
        - type: object
          required: [regex]
          properties:
            regex: {type: string}
        - type: object
          required: [present]
          properties:
            present: {type: boolean}
    BodyMatcher:
      description: Exactly one operator.
      oneOf:
        - type: object
          required: [equals]
          properties:
            equals: {type: string}
        - type: object
          required: [contains]
          properties:
            contains: {type: string}
        - type: object
          required: [regex]
          properties:
            regex: {type: string}
        - type: object
          required: [jsonPath]
          properties:
            jsonPath:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                equals:
                  description: Any JSON value; without it, the path only has to exist.
        - type: object
          required: [equalToJson]
          properties:
            equalToJson:
              description: A JSON value, or a string holding one.
...
    Fault:
      description: A network failure instead of the response.
      enum:
        - connection_reset
        - empty_response
        - malformed_chunk
        - random_data_then_close
        - hang
        - stall_after_headers
        - truncated_body
...
    SearchRequest:
      type: object
      properties:
        text:
          type: [string, 'null']
          maxLength: 200
          description: >-
            Case-insensitive substring of method, url, ip, header or query names and values, body, or names and
            values of the decrypted attribute (`decrypted`, at any level). The MCP `search_requests` tool never
            searches the decrypted value.
        match:
          oneOf:
            - $ref: '#/components/schemas/RuleMatch'
            - type: 'null'
        outcome:
          description: The recorded outcome.
          oneOf:
            - type: object
              required: [type, rule]
              properties:
                type:
                  enum: [rule, near_miss]
                rule:
                  type: string
                  format: uuid
            - type: object
              required: [type]
              properties:
                type:
                  const: default
            - type: 'null'
        signature_reason:
          type: [string, 'null']
          minLength: 1
          maxLength: 200
          description: Invalid signature with this reason (ignoring a trailing parenthesis).
        schema_path:
          type: [string, 'null']
          maxLength: 1000
          description: Invalid schema with an error at this JSON Pointer.
        sorting:
          enum: [newest, oldest, null]
          default: newest
        page:
          type: [integer, 'null']
          minimum: 1
          default: 1
        per_page:
          type: [integer, 'null']
          minimum: 1
          maximum: 100
          default: 50
...